@@ -2,7 +2,8 @@ package com.mycompany.invoise.core.entity.customer;
 
 import com.mycompany.invoise.core.entity.customer.Address;
 
-import javax.persistence.*;
+
+import jakarta.persistence.*;
 
 @Entity
 public class Customer {
@@ -10,11 +11,11 @@ public class Customer {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    @Column(nullable = false,length = 50)
+    @Column(nullable = false, length = 50)
     private String name;
 
-    @OneToOne(fetch = FetchType.LAZY, cascade = CascadeType.ALL,orphanRemoval = true,optional = false)
-    @JoinColumn(name="ID_ADDRESS")
+    @OneToOne(fetch = FetchType.LAZY, cascade = CascadeType.ALL, orphanRemoval = true, optional = false)
+    @JoinColumn(name = "ID_ADDRESS")
     private Address address;
 
     public Customer(String name) {
