@@ -1,7 +1,7 @@
 package com.mycompany.invoise.core.entity.product;
 
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 
 @Entity
 public class Category {
@@ -9,7 +9,7 @@ public class Category {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    @Column(nullable = false,length = 50)
+    @Column(nullable = false, length = 50)
     private String name;
 
     public Category(String name) {
