@@ -2,8 +2,9 @@ package com.mycompany.invoise.core.entity.invoice;
 
 
 import com.mycompany.invoise.core.entity.customer.Customer;
+import jakarta.persistence.*;
 
-import javax.persistence.*;
+
 import java.util.ArrayList;
 import java.util.List;
 
