@@ -4,7 +4,7 @@ import com.mycompany.invoise.core.entity.customer.Customer;
 import org.springframework.data.repository.CrudRepository;
 
 
-public interface CustomerRepositoryInterface extends CrudRepository<Customer,Long> {
+public interface CustomerRepositoryInterface extends CrudRepository<Customer, Long> {
     /*Invoice create(Invoice invoice);
     List<Invoice> list();
     Invoice getById(String number);*/
