@@ -25,8 +25,8 @@ public class InvoiceServicePrefix implements InvoiceServiceInterface {
         this.invoiceRepository = invoiceRepository;
     }
 
-    public Invoice createInvoice(Invoice invoice){
-        invoice.setNumber(prefix+(++lastNumber));
+    public Invoice createInvoice(Invoice invoice) {
+        invoice.setNumber(prefix + (++lastNumber));
         invoiceRepository.save(invoice);
         return invoice;
     }
